requirements.txt
google-cloud-bigquery>=3.20
requests>=2.31
pytz>=2024.1
